requirements.txt
streamlit
python-dotenv
Pillow
PyMuPDF
sentence-transformers
openai
faiss-cpu